change whitespace style for rules (details below)
I think I prefer to put the goal that a rule applies to on a separate
line, so that it is aligned with the subgoals. This looks much closer to
how a rule looks is written on paper, and is easier to read IMO.

Feel free to revert if this is especially unidiomatic.