IDEA: make coercions only deal with one layer, no recursion
The idea is that by "abstraction" there is no way to construct a term
where the non-wellformed part is anywhere other than the outer layer of
syntax. Basically, within the logical-framework module we have to
maintain this invariant, but outside that module, we only ever need to
work with a single layer of wellformedness.

The result is that we can *always* call these coercions everywhere,
without needing to remember if the term has been "fixed" yet.

*******EXCEPTION****** The one way to break this abstraction is using
the primitive abstract/instantiate stuff from the locally-nameless
library. I think that somehow we should make this abstract so that users
of the LF can never do this.